Refactoring the subscription mechanism.
