+++ CMakeLists.txt
@@ -40,6 +40,7 @@ set(
     src/rws_poco_client.cpp
     src/rws_rapid.cpp
     src/rws_state_machine_interface.cpp
+    src/rws_subscription.cpp
 )
 
 add_library(${PROJECT_NAME} ${SRC_FILES})
